Generación de token y protección de endpoints

+++ src/main/java/com/procesos/negocio/karen/controllers/UsuarioControllers.java
@@ -3,6 +3,7 @@ package com.procesos.negocio.karen.controllers;
 import aj.org.objectweb.asm.Opcodes;
 import com.procesos.negocio.karen.models.Usuario;
 import com.procesos.negocio.karen.services.UsuarioService;
+import com.procesos.negocio.karen.utils.JWTUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -20,10 +21,17 @@ public class UsuarioControllers {
 
     @Autowired
     private UsuarioService usuarioService;
+    @Autowired
+    private JWTUtil jwtUtil;
 
-    @GetMapping(value = "/usuario/{id}")
+    @GetMapping(value = "/usuario/id}")
 
-    public ResponseEntity getUsuario(@PathVariable Long id) {
+    public ResponseEntity getUsuario(@PathVariable Long id,
+                                     @RequestHeader(value = "Authorization") String token) {
+        if(jwtUtil.getKey(token)==null)
+        {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no valido");
+      }
         return usuarioService.getUserById(id);
     }
 
@@ -35,34 +43,66 @@ public class UsuarioControllers {
 
 
     @GetMapping("/usuarios")
-    public ResponseEntity listarUsuarios() {
+    public ResponseEntity listarUsuarios(
+            @RequestHeader(value = "Authorization") String token) {
+        if(jwtUtil.getKey(token)==null)
+        {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no valido");
+        }
+
         return usuarioService.allUsers();
     }
 
     @GetMapping("/usuario/{nombre}/{apellidos}")
-    public ResponseEntity ListarPorNombreApellidos(@PathVariable String nombre, @PathVariable String apellidos) {
+    public ResponseEntity ListarPorNombreApellidos(@PathVariable String nombre, @PathVariable String apellidos,
+                                                   @RequestHeader(value = "Authorization") String token) {
+        if(jwtUtil.getKey(token)==null)
+        {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no valido");
+        }
+
         return usuarioService.allUsersByNameAndLastName(nombre,apellidos);
     }
 
     @GetMapping("/usuarios/nombre/{nombre}")
 
-    public ResponseEntity ListarPorNombre(@PathVariable String nombre) {
+    public ResponseEntity ListarPorNombre(@PathVariable String nombre,
+                                          @RequestHeader(value = "Authorization") String token) {
+        if(jwtUtil.getKey(token)==null)
+        {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no valido");
+        }
         return usuarioService.allUsersByName(nombre);
     }
 
     @GetMapping("/usuarios/apellidos/{apellidos}")
 
-    public ResponseEntity ListarPorApellidos(@PathVariable String apellidos) {
+    public ResponseEntity ListarPorApellidos(@PathVariable String apellidos,
+                                             @RequestHeader(value = "Authorization") String token) {
+        if(jwtUtil.getKey(token)==null)
+        {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no valido");
+        }
         return usuarioService.allUsersByLastName(apellidos);
     }
 
     @PutMapping("/usuario/{id}")
-    public ResponseEntity editarUsuario(@PathVariable Long id, @Valid @RequestBody Usuario usuario) {
+    public ResponseEntity editarUsuario(@PathVariable Long id, @Valid @RequestBody Usuario usuario,
+                                        @RequestHeader(value = "Authorization") String token) {
+        if(jwtUtil.getKey(token)==null)
+        {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no valido");
+        }
         return usuarioService.editUser(id,usuario);
     }
 
     @DeleteMapping("/usuario/{id}")
-    public ResponseEntity eliminarUsuario(@PathVariable Long id) {
+    public ResponseEntity eliminarUsuario(@PathVariable Long id,
+                                          @RequestHeader(value = "Authorization") String token) {
+        if(jwtUtil.getKey(token)==null)
+        {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Token no valido");
+        }
         return usuarioService.deleteUser(id);
     }
